# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_api.CreateLedgerRequest:
    properties:
      currency:
        example: CNY
        type: string
      name:
        example: My Ledger
        maxLength: 100
        type: string
      type:
        enum:
        - personal
        - shared
        example: personal
        type: string
    required:
    - name
    type: object
...
  internal_api.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      error:
        example: bad request
        type: string
      message:
        example: Invalid input
        type: string
    type: object
  internal_api.LedgerResponse:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      currency:
        example: CNY
        type: string
      id:
        example: 1
        type: integer
      name:
        example: My Ledger
        type: string
      type:
        example: personal
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
...
  internal_api.UpdateLedgerRequest:
    properties:
      currency:
        example: USD
        type: string
      name:
        example: Updated Ledger
        maxLength: 100
        type: string
      type:
        enum:
        - personal
        - shared
        example: shared
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: BeeCount Cloud API 是一个用于管理个人财务账本的云服务 API，提供账本、交易、分类、标签等功能。
  license:
    name: AGPL 3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  termsOfService: http://swagger.io/terms/
  title: BeeCount Cloud API
  version: "1.0"
paths:
...
  /api/v1/ledgers:
    get:
      consumes:
      - application/json
      description: 返回当前用户拥有的所有账本列表
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            items:
              $ref: '#/definitions/internal_api.LedgerResponse'
            type: array
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: 获取账本列表
      tags:
      - 账本
    post:
      consumes:
      - application/json
      description: 为当前用户创建新的账本
      parameters:
      - description: 账本信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.CreateLedgerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            $ref: '#/definitions/internal_api.LedgerResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: 创建账本
      tags:
      - 账本
  /api/v1/ledgers/{ledger_id}:
    delete:
      consumes:
      - application/json
      description: 删除指定账本，只有账本所有者可以删除
      parameters:
      - description: 账本ID
        in: path
        name: ledger_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: 删除成功
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: 账本不存在
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: 删除账本
      tags:
      - 账本
    get:
      consumes:
      - application/json
      description: 根据账本ID返回账本详细信息
      parameters:
      - description: 账本ID
        in: path
        name: ledger_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/internal_api.LedgerResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: 账本不存在
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: 获取账本详情
      tags:
      - 账本
    put:
      consumes:
      - application/json
      description: 更新指定账本的信息，只有账本所有者可以更新
      parameters:
      - description: 账本ID
        in: path
        name: ledger_id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.UpdateLedgerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/internal_api.LedgerResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: 账本不存在
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: 更新账本
      tags:
      - 账本
  /api/v1/ledgers/{ledger_id}/sync/download:
    get:
      consumes:
      - application/json
      description: 返回账本数据的JSON格式，用于客户端同步
      parameters:
      - description: 账本ID
        in: path
        name: ledger_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 下载成功，返回JSON数据
          schema:
            type: string
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: 账本不存在
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: 下载账本数据
      tags:
      - 同步
  /api/v1/ledgers/{ledger_id}/sync/status:
    get:
      consumes:
      - application/json
      description: 返回账本的同步状态信息，包括交易数量、最后更新时间等
      parameters:
      - description: 账本ID
        in: path
        name: ledger_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: 账本不存在
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: 获取同步状态
      tags:
      - 同步
  /api/v1/ledgers/{ledger_id}/sync/upload:
    post:
      consumes:
      - application/json
      description: 将账本数据导出为JSON格式，用于客户端同步
      parameters:
      - description: 账本ID
        in: path
        name: ledger_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功，返回JSON数据
          schema:
            type: string
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: 上传账本数据
      tags:
      - 同步
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"